Refactor move fixture folder a level above
